refactor popup into separate file
